+++ packages/apptive_grid_form/lib/apptive_grid_form.dart
@@ -31,6 +31,9 @@ class ApptiveGridForm extends StatefulWidget {
     this.onFormLoaded,
     this.onActionSuccess,
     this.onError,
+    this.scrollController,
+    this.buttonAlignment = Alignment.center,
+    this.buttonLabel,
   }) : super(key: key);
 
   /// [FormUri] of the Form to display
@@ -44,10 +47,10 @@ class ApptiveGridForm extends StatefulWidget {
   final TextStyle? titleStyle;
 
   /// Padding of the Items in the Form. If no Padding is provided a EdgeInsets.all(8.0) will be used.
-  final EdgeInsets? contentPadding;
+  final EdgeInsetsGeometry? contentPadding;
 
   /// Padding for the title. If no Padding is provided the [contentPadding] is used
-  final EdgeInsets? titlePadding;
+  final EdgeInsetsGeometry? titlePadding;
 
   /// Flag to hide the form title, default is false
   final bool hideTitle;
@@ -79,6 +82,16 @@ class ApptiveGridForm extends StatefulWidget {
   /// This functionality can be used to do a custom Widget or Transition
   final Future<bool> Function(dynamic)? onError;
 
+  /// Optional ScrollController for the Form
+  final ScrollController? scrollController;
+
+  /// Alignment of the Send Button
+  final Alignment buttonAlignment;
+
+  /// Label of the Button to submit a form.
+  /// Defaults to a localized version of `Send`
+  final String? buttonLabel;
+
   @override
   ApptiveGridFormState createState() => ApptiveGridFormState();
 }
@@ -99,7 +112,7 @@ class ApptiveGridFormState extends State<ApptiveGridForm> {
   void didUpdateWidget(covariant ApptiveGridForm oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.formUri != oldWidget.formUri) {
-      _loadForm();
+      loadForm();
     }
   }
 
@@ -107,7 +120,7 @@ class ApptiveGridFormState extends State<ApptiveGridForm> {
   void didChangeDependencies() {
     super.didChangeDependencies();
     _client = ApptiveGrid.getClient(context);
-    _loadForm();
+    loadForm();
   }
 
   @override
@@ -122,11 +135,23 @@ class ApptiveGridFormState extends State<ApptiveGridForm> {
       hideTitle: widget.hideTitle,
       onActionSuccess: widget.onActionSuccess,
       onError: widget.onError,
-      triggerReload: _loadForm,
+      triggerReload: () => loadForm(resetData: false),
+      scrollController: widget.scrollController,
+      buttonAlignment: widget.buttonAlignment,
+      buttonLabel: widget.buttonLabel,
     );
   }
 
-  void _loadForm() {
+  /// Loads the form
+  ///
+  /// If [resetData] is `true` it will call setState with [_formData] = `null`
+  /// Useful if a manual reload of the form is required
+  void loadForm({bool resetData = false}) {
+    if (resetData) {
+      setState(() {
+        _formData = null;
+      });
+    }
     _client.loadForm(formUri: widget.formUri).then((value) {
       if (widget.onFormLoaded != null) {
         widget.onFormLoaded!(value);
@@ -164,6 +189,9 @@ class ApptiveGridFormData extends StatefulWidget {
     this.onActionSuccess,
     this.onError,
     this.triggerReload,
+    this.scrollController,
+    this.buttonAlignment = Alignment.center,
+    this.buttonLabel,
   }) : super(key: key);
 
   /// [FormData] that should be displayed
@@ -176,10 +204,10 @@ class ApptiveGridFormData extends StatefulWidget {
   final TextStyle? titleStyle;
 
   /// Padding of the Items in the Form. If no Padding is provided a EdgeInsets.all(8.0) will be used.
-  final EdgeInsets? contentPadding;
+  final EdgeInsetsGeometry? contentPadding;
 
   /// Padding for the title. If no Padding is provided the [contentPadding] is used
-  final EdgeInsets? titlePadding;
+  final EdgeInsetsGeometry? titlePadding;
 
   /// Flag to hide the form title, default is false
   final bool hideTitle;
@@ -198,6 +226,16 @@ class ApptiveGridFormData extends StatefulWidget {
 
   /// Will be called when [formData] should be reloaded
   final void Function()? triggerReload;
+
+  /// Optional ScrollController for the Form
+  final ScrollController? scrollController;
+
+  /// Alignment of the Send Button
+  final Alignment buttonAlignment;
+
+  /// Label of the Button to submit a form.
+  /// Defaults to a localized version of `Send`
+  final String? buttonLabel;
 
   @override
   ApptiveGridFormDataState createState() => ApptiveGridFormDataState();
@@ -301,6 +339,7 @@ class ApptiveGridFormDataState extends State<ApptiveGridFormData> {
       child: Form(
         key: _formKey,
         child: ListView.builder(
+          controller: widget.scrollController,
           itemCount:
               1 + (data.components?.length ?? 0) + (submitLink != null ? 1 : 0),
           itemBuilder: (context, index) {
@@ -340,20 +379,28 @@ class ApptiveGridFormDataState extends State<ApptiveGridFormData> {
             } else {
               return Padding(
                 padding: widget.contentPadding ?? _defaultPadding,
-                child: Builder(
-                  builder: (_) {
-                    if (_actionsInProgress.contains(submitLink)) {
-                      return const Center(
-                        child: CircularProgressIndicator.adaptive(),
-                      );
-                    } else {
-                      return ActionButton(
-                        action: submitLink!,
-                        onPressed: _submitForm,
-                        child: Text(localization.actionSend),
-                      );
-                    }
-                  },
+                child: Align(
+                  alignment: widget.buttonAlignment,
+                  child: Builder(
+                    builder: (_) {
+                      if (_actionsInProgress.contains(submitLink)) {
+                        return const TextButton(
+                          onPressed: null,
+                          child: Center(
+                            child: CircularProgressIndicator.adaptive(),
+                          ),
+                        );
+                      } else {
+                        return ActionButton(
+                          action: submitLink!,
+                          onPressed: _submitForm,
+                          child: Text(
+                            widget.buttonLabel ?? localization.actionSend,
+                          ),
+                        );
+                      }
+                    },
+                  ),
                 ),
               );
             }
@@ -366,6 +413,7 @@ class ApptiveGridFormDataState extends State<ApptiveGridFormData> {
   Widget _buildSuccess(BuildContext context) {
     final localization = ApptiveGridLocalization.of(context)!;
     return ListView(
+      controller: widget.scrollController,
       padding: const EdgeInsets.all(32.0),
       children: [
         AspectRatio(
@@ -398,6 +446,7 @@ class ApptiveGridFormDataState extends State<ApptiveGridFormData> {
   Widget _buildSaved(BuildContext context) {
     final localization = ApptiveGridLocalization.of(context)!;
     return ListView(
+      controller: widget.scrollController,
       padding: const EdgeInsets.all(32.0),
       children: [
         AspectRatio(
@@ -429,6 +478,7 @@ class ApptiveGridFormDataState extends State<ApptiveGridFormData> {
     final localization = ApptiveGridLocalization.of(context)!;
     final theme = Theme.of(context);
     return ListView(
+      controller: widget.scrollController,
       padding: const EdgeInsets.all(32.0),
       children: [
         AspectRatio(